+++ CHANGELOG.md
@@ -1183,6 +1183,7 @@ TV-9.4.0.20220825
     - #066463 - The 'Term & Conditions and 'Privacy Policies' links are not clickable on the signup page
     - #066465 - The user email address appears prefilled on the sign-in with the phone number field.
     - #066409 - The page heading is missing on some pages.
+    - #066438 - Suggestion: To add the sign-in button on the registration successful page
 ## UAT points:
 
     - #Admin dashboard sales stats > set tabs priority, Move Affiliate tab to the last